# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "为医院信息系统提供标准化的医保结算接口服务：参保人身份核验、费用上传、预结算、正式结算、结算单查询与日终对账。业务接口需 API Key（X-API-Key）与 JWT（Authorization: Bearer）双认证。",
        "title": "医保智能结算 API 网关",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:19935",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/settlement-adjustments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "settlement-adjustments"
                ],
                "summary": "差额补退单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "调用方 ID",
                        "name": "client_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态(pending_review/approved/rejected)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/settlement-adjustments/{adjustment_no}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "settlement-adjustments"
                ],
                "summary": "差额补退单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "差额补退单号",
                        "name": "adjustment_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/settlement-adjustments/{adjustment_no}/review": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "复核通过后订单标记已补退并生成负向记录冲减差额；驳回保留原因且可再次申请；重复/并发复核只成功一次",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settlement-adjustments"
                ],
                "summary": "差额补退复核",
                "parameters": [
                    {
                        "type": "string",
                        "description": "差额补退单号",
                        "name": "adjustment_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "复核结论（approved + reason）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReviewAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/settlements/{settlement_no}/adjustments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "settlement-adjustments"
                ],
                "summary": "结算单差额补退记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "结算单号",
                        "name": "settlement_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "已结算订单提交目标医保支付额与原因后生成待复核差额单；同一订单只允许一条待复核记录，目标金额小于零或不小于原支付额时拒绝",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settlement-adjustments"
                ],
                "summary": "结算差额补退申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "结算单号",
                        "name": "settlement_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "目标医保支付额与原因",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateAdjustmentRequest": {
            "type": "object",
            "required": [
                "reason",
                "target_pay_amount"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 500
                },
                "target_pay_amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.ReviewAdjustmentRequest": {
            "type": "object",
            "properties": {
                "approved": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "util.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}